Get by reference and const-ref
Have to use a dedicated "reference type" so that we can wrap the
UnitChecking around it. This type decays to a "normal" type
Allow 3 cases depending on constness and L/R-value-ness. This
matches behaviour of regular types
Adds various helpers to enable what we're doing
Adds tests, including some to ensure coverage